% File src/library/bigsplines/man/grid-internal.Rd
% Part of the R bigsplines package, 
% Nathaniel E. Helwig <helwig@umn.edu>

\name{bigsplines-internal}

%% List of Internal Functions Called
\alias{gcvcss}
\alias{gcvgss}
\alias{gcvssa}
\alias{gcvssg}
\alias{gcvssp}
\alias{lamcoef}
\alias{lamcoefg}
\alias{lamloop}
\alias{lamloopg}
\alias{makerkm}
\alias{nbmle}
\alias{pdsXty}
\alias{pinvsm}
\alias{postvar}
\alias{rkron}
\alias{smartssa}
\alias{smartssg}
\alias{smartssp}
\alias{ssawork}
\alias{ssadpm}
\alias{ssgwork}
\alias{sspwork}
\alias{sspdpm}
\alias{tcprod}
\alias{unifqsum}
\alias{unifqsumg}


\title{Internal functions for big splines package}
\description{
  Internal functions for big splines package.
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }

